refactor: move test hooks
