+++ agent/src/main/java/com/appland/appmap/process/hooks/test/TestSupport.java
@@ -1,0 +1,6 @@
+package com.appland.appmap.process.hooks.test;
+
+class TestSupport {
+  static final String TEST_RECORDER_TYPE = "tests";
+
+}
